Move hook setup to separate file to declutter BingoSync.cs for future outwards interface
