+++ BingoSync/BingoSync.cs
@@ -19,149 +19,11 @@ namespace BingoSync
         public static string version = "1.3.0.0";
         public override string GetVersion() => version;
 
-        const bool Debug = false;
-
         public static ModSettings modSettings { get; set; } = new ModSettings();
 
         public override void Initialize()
         {
-            // Check bingo objectives every frame
-            ModHooks.HeroUpdateHook += HeroUpdate;
-
-            // General
-            ModHooks.SetPlayerBoolHook += UpdateBoolInternal;
-            ModHooks.SetPlayerIntHook += UpdateIntInternal;
-
-            // GeoSpent
-            On.GeoCounter.TakeGeo += GeoSpent.UpdateGeoSpent;
-            On.GeoCounter.Update += GeoSpent.UpdateGeoText;
-
-            // Tolls
-            On.GeoCounter.TakeGeo += Tolls.UpdateTolls;
-
-            // Grubs
-            ModHooks.SetPlayerIntHook += Grubs.CheckIfGrubWasSaved;
-
-            // Myla
-            ModHooks.SetPlayerIntHook += Myla.CheckIfMylaWasKilled;
-
-            // Revek
-            On.HeroController.NailParry += Revek.CheckParry;
-            On.HeroController.EnterScene += Revek.EnterRoom;
-
-            // Lifts
-            ModHooks.SetPlayerBoolHook += Lifts.CheckIfLiftWasUsed;
-
-            // Jiji
-            ModHooks.SetPlayerBoolHook += Jiji.CheckIfKilledShadeInJijis;
-
-            // Dialogue
-            On.DialogueBox.StartConversation += Dialogue.StartConversation;
-
-            // Hot Springs
-            ModHooks.SetPlayerIntHook += HotSprings.CheckBath;
-
-            // Hive Shard
-            ModHooks.SetPlayerBoolHook += HiveShard.CheckIfHiveShardWasCollected;
-
-            // Tram
-            ModHooks.SetPlayerIntHook += Tram.CheckIfStationWasVisited;
-
-            // Unique Enemies
-            ModHooks.OnReceiveDeathEventHook += UniqueEnemies.CheckIfUniqueEnemyWasKilled;
-            On.ScuttlerControl.Hit += UniqueEnemies.HitLightseed;
-            On.HealthManager.TakeDamage += UniqueEnemies.KillGulkaWithSpikeBall;
-            On.HealthManager.TakeDamage += UniqueEnemies.OomasKilledWithMinions_GlowingWomb;
-            On.SetHP.OnEnter += UniqueEnemies.OomasKilledWithMinions_Weaversong_Grimmchild;
-
-            // Giant Geo Egg
-            On.PlayMakerFSM.OnEnable += GiantGeoEgg.CreateGiantGeoRockTrigger;
-
-            // Marissa
-            On.PlayMakerFSM.OnEnable += Marissa.CreateMarissaKilledTrigger;
-
-            // Stag
-            On.PlayMakerFSM.OnEnable += Stag.CreateStagTravelTrigger;
-
-            // BreakableFloors
-            On.PlayMakerFSM.OnEnable += BreakableFloors.CreateBreakableFloorsTrigger;
-
-            // Oro Training Dummy
-            On.PlayMakerFSM.OnEnable += OroTrainingDummy.CreateOroTrainingDummyTrigger;
-
-            // Millibelle
-            On.PlayMakerFSM.OnEnable += Millibelle.CreateMillibelleHitTrigger;
-
-            // Chests
-            On.PlayMakerFSM.OnEnable += Chests.CreateChestOpenTrigger;
-
-            // Switches
-            On.PlayMakerFSM.OnEnable += Switches.CreateSwitchOpenTrigger;
-
-            // Benches
-            On.PlayMakerFSM.OnEnable += Benches.CreateBenchTrigger;
-
-            // Tiso
-            On.PlayMakerFSM.OnEnable += Tiso.CreateTisoShieldTrigger;
-
-            // Telescope
-            On.PlayMakerFSM.OnEnable += Telescope.CreateTelescopeTrigger;
-
-            // Shade Gates
-            On.PlayMakerFSM.OnEnable += ShadeGates.CreateShadeGateTrigger;
-
-            // Dream Nail Dialogue
-            On.PlayMakerFSM.OnEnable += DreamNailDialogue.CreateDreamNailDialogueTrigger;
-
-            // Lore Tablets
-            On.PlayMakerFSM.OnEnable += LoreTablets.CreateLoreTabletTrigger;
-
-            // Nail Arts
-            On.PlayMakerFSM.OnEnable += NailArts.CreateNailArtsTrigger;
-
-            // Spa Gladiator
-            On.PlayMakerFSM.OnEnable += SpaGladiator.CreateSplashedTrigger;
-
-            // Eternal Ordeal
-            On.PlayMakerFSM.OnEnable += EternalOrdeal.CreateCounterTrigger;
-
-            // Void Pool
-            On.PlayMakerFSM.OnEnable += VoidPool.CreateVoidPoolTrigger;
-
-            // Shinies
-            On.PlayMakerFSM.OnEnable += Shinies.CreateTrinketTrigger;
-
-            // City Gate
-            On.PlayMakerFSM.OnEnable += CityGate.CreateCityGateOpenedTrigger;
-
-            // Scenes
-            On.HeroController.EnterScene += Scenes.EnterRoom;
-            
-            // Charms
-            ModHooks.SetPlayerBoolHook += Charms.CheckEquippedCharms;
-
-            // Bow
-            ModHooks.HeroUpdateHook += Bow.BowToNPC;
-
-            // NailHit
-            ModHooks.SlashHitHook += NailHit.ProcessNailHit;
-
-            // Rando
-            AbstractItem.AfterGiveGlobal += Checks.AfterGiveItem;
-            AbstractPlacement.OnVisitStateChangedGlobal += Checks.PlacementStateChange;
-
-            // Menu
-            On.UIManager.ContinueGame += ContinueGame;
-            On.UIManager.StartNewGame += StartNewGame;
-            On.UIManager.FadeInCanvasGroup += FadeIn;
-            On.UIManager.FadeOutCanvasGroup += FadeOut;
-
-            var _hook = new ILHook
-            (
-                typeof(DreamPlant).GetMethod("CheckOrbs", BindingFlags.NonPublic | BindingFlags.Instance).GetStateMachineTarget(),
-                DreamTrees.TrackDreamTrees
-            );
-
+            Hooks.Setup();
             RetryHelper.Setup(Log);
             MenuUI.Setup();
             BingoSyncClient.Setup(Log);
@@ -177,72 +39,6 @@ namespace BingoSync
         private void HideMenu()
         {
             MenuUI.SetVisible(false);
-        }
-
-        private IEnumerator FadeOut(On.UIManager.orig_FadeOutCanvasGroup orig, UIManager self, CanvasGroup cg)
-        {
-            if (cg.name == "MainMenuScreen")
-            {
-                HideMenu();
-            }
-            return orig(self, cg);
-        }
-
-        private IEnumerator FadeIn(On.UIManager.orig_FadeInCanvasGroup orig, UIManager self, CanvasGroup cg)
-        {
-            if (cg.name == "MainMenuScreen")
-            {
-                ShowMenu();
-            }
-            return orig(self, cg);
-        }
-
-        private void ContinueGame(On.UIManager.orig_ContinueGame orig, UIManager self)
-        {
-            HideMenu();
-            ConfigureBingoSyncOnGameStart();
-            Task.Run(() => {
-                Checks.GetRandomizedPlacements();
-            });
-            orig(self);
-        }
-
-        private void StartNewGame(On.UIManager.orig_StartNewGame orig, UIManager self, bool permaDeath, bool bossRush)
-        {
-            HideMenu();
-            ConfigureBingoSyncOnGameStart();
-            Task.Run(() => {
-                Checks.GetRandomizedPlacements();
-            });
-            orig(self, permaDeath, bossRush);
-        }
-
-        private void ConfigureBingoSyncOnGameStart()
-        {
-            if (!modSettings.RevealCardOnGameStart) return;
-            BingoSyncClient.RevealCard();
-        }
-
-        private void HeroUpdate()
-        {
-            BingoTracker.ProcessBingo();
-        }
-
-        private bool UpdateBoolInternal(string name, bool orig)
-        {
-            BingoTracker.UpdateBoolean(name, orig);
-            if (Debug)
-                Log($"bool: {name} {orig} {GameManager.instance.GetSceneNameString()}");
-            return orig;
-        }
-
-        private int UpdateIntInternal(string name, int current)
-        {
-            var previous = PlayerData.instance.GetIntInternal(name);
-            BingoTracker.UpdateInteger(name, previous, current);
-            if (Debug)
-                Log($"int: {name} {previous} {current} {GameManager.instance.GetSceneNameString()}");
-            return current;
         }
 
         public void OnLoadLocal(Settings.SaveSettings s)
